fixed numerous compiler warnings on OSX

+++ Simperium/NSData+Simperium.m
@@ -32,7 +32,7 @@ static const short _base64DecodingTable[256] = {
 
 + (NSData *)decodeBase64WithString:(NSString *)strBase64 {
     const char * objPointer = [strBase64 cStringUsingEncoding:NSASCIIStringEncoding];
-    int intLength = strlen(objPointer);
+    int intLength = (int)strlen(objPointer);
     int intCurrent;
     int i = 0, j = 0, k;
